requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
openpyxl==3.1.2
google-cloud-vision==3.4.4